backend_setup/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
PyJWT==2.8.0
flask-jwt-extended==4.5.3